pyproject.toml
[project]
name = "py5gmeta_akkodis"
version = "0.0.1"

authors = [
  { name="Djibrilla Amadou Kountche", email="5gmeta@akkodis.com"}
]

description = "A Python Library for interaction with the 5GMETA Platform"
readme = " README.md"

requires-python = ">3.10"
dependencies = [
  "python-qpid-proton",
  "confluent-kafka",
  "kafka-python",
  "numpy",
  "requests",
  "confluent-kafka",
  "avro",
  "fastavro",
  "avro_to_python",
  "pygeotile",
  "matplotlib",
  "folium",
  "sqlalchemy",
  "python-keycloak",
  "prometheus-client"
]

classifiers = [
     "Programming Language :: Python :: 3",
     "Operating System :: OS Independent",
]

license = "EUPL-1.2"
license-files = ["LICEN[CS]E*"]


[project.urls]
Homepage = "https://github.com/Akkodis/5gmeta"
Isssues  = "https://github.com/Akkodis/5gmeta/issues"


[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"
